api/requirements.txt
Django==5.2.5
dj-database-url==3.0.1
whitenoise==6.9.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
